More Test Cases for Endpoint Testing
