+++ api/src/tests/testvalues.txt
@@ -1,0 +1,10 @@
+
+1234
+hellothere
+oA9B23B2235ZzxDf4Hg92Dg%200$528fG3pL892NNb2M8cc0mljsFv732m9c8s8gSddsBc73
+-2342334234
+-5BILLIONmiles
+number
+testing1234 iririririririririririririoeoedksmdiedmsoekrewicdiesdslkemdlsiemdlskmdlsek
+jimmy
+scccceeeeeeeeeeeeeeee
